+++ reddit_api_kernel/submit_text.mli
@@ -2,6 +2,6 @@ open! Core
 
 type t [@@deriving sexp]
 
-val of_json : Json.t -> t
-val to_json : t -> Json.t
+include Json_object.S with type t := t
+
 val submit_text : [ `markdown | `HTML ] -> t -> string