+++ test/test_modmail.ml
@@ -18,44 +18,41 @@ let%expect_test "create_modmail_conversation" =
       [%expect
         {|
         ((conversation
-          (O
-           ((isAuto (Bool false))
-            (objIds (A ((O ((id (String osvgj)) (key (String messages)))))))
-            (isRepliable (Bool true)) (lastUserUpdate Null) (isInternal (Bool false))
+          (Object
+           ((isAuto False)
+            (objIds (Array ((Object ((id (String osvgj)) (key (String messages)))))))
+            (isRepliable True) (lastUserUpdate Null) (isInternal False)
             (lastModUpdate (String 2020-07-26T21:18:43.146061+00:00))
             (lastUpdated (String 2020-07-26T21:18:43.146061+00:00))
             (authors
-             (A
-              ((O
-                ((isMod (Bool true)) (isAdmin (Bool false))
-                 (name (String L72_Elite_Kraken)) (isOp (Bool true))
-                 (isParticipant (Bool false)) (isHidden (Bool false))
-                 (id (Float 71814082)) (isDeleted (Bool false)))))))
+             (Array
+              ((Object
+                ((isMod True) (isAdmin False) (name (String L72_Elite_Kraken))
+                 (isOp True) (isParticipant False) (isHidden False)
+                 (id (Number 71814082)) (isDeleted False))))))
             (owner
-             (O
+             (Object
               ((displayName (String ThirdRealm)) (type (String subreddit))
                (id (String t5_390u2)))))
-            (id (String fsv44)) (isHighlighted (Bool false))
-            (subject (String "Test subject")) (participant (O ())) (state (Float 0))
-            (lastUnread Null) (numMessages (Float 1)))))
+            (id (String fsv44)) (isHighlighted False)
+            (subject (String "Test subject")) (participant (Object ()))
+            (state (Number 0)) (lastUnread Null) (numMessages (Number 1)))))
          (messages
-          (O
+          (Object
            ((osvgj
-             (O
+             (Object
               ((body
                 (String
                   "<!-- SC_OFF --><div class=\"md\"><p>Test body</p>\
                  \n</div><!-- SC_ON -->"))
                (author
-                (O
-                 ((isMod (Bool true)) (isAdmin (Bool false))
-                  (name (String L72_Elite_Kraken)) (isOp (Bool true))
-                  (isParticipant (Bool false)) (isHidden (Bool false))
-                  (id (Float 71814082)) (isDeleted (Bool false)))))
-               (isInternal (Bool false))
-               (date (String 2020-07-26T21:18:43.146061+00:00))
+                (Object
+                 ((isMod True) (isAdmin False) (name (String L72_Elite_Kraken))
+                  (isOp True) (isParticipant False) (isHidden False)
+                  (id (Number 71814082)) (isDeleted False))))
+               (isInternal False) (date (String 2020-07-26T21:18:43.146061+00:00))
                (bodyMarkdown (String "Test body")) (id (String osvgj))))))))
-         (modActions (O ()))) |}];
+         (modActions (Object ()))) |}];
       return ())
 ;;